server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pytest
httpx
requests
watchdog
dropbox>=12.0.0
